Test: check how expensive attacker can make finalizeExit and addExit
Honest actors start to drain exits queue as soon as attack is detected.
This gives attacker at most two weeks to spam exit queue (this is how
long fresh UTXO has to stay in the queue).

Other costs are not evaluated (mappings inserts and deletes) since they
are independent of size of the queue.