fix: process ife without piggybacked would fail
This commit fixes the bug that an in-flight exit without piggybacked would
cause "processExits" to fail. This is because in in-flight exits, they are
put into priority queue only when they are piggybacked. And "processExits"
would crash on an empty queue.

Fix the issue by enabling "processExits" on an empty queue.

- issue: https://github.com/omisego/plasma-contracts/issues/73