fix: inputs of deposit not checked
Current contract does not check the input of deposit tx
should be empty. As a result, currently user can fake any input
data to the deposit tx. This may lead to some unexpected behaviour,
eg. making an IFE exit_id not unique by mimicing another existing
transaction with the same input. We did not separate deposit exit id
from regular tx for IFE because it is assumed that deposit tx cannot IFE
due to sum(inputs) would be 0 < sum(outputs). However, if fake input
is possible, then the assumption would be wrong.

This commit add check to deposit tx inputs on `blknum` to be 0,
which should be enough as block start with 1.

issue: https://github.com/omisego/plasma-contracts/issues/88