style: use event add comments for test code
Previously the check of "onlyXXX" is done by checking
overriding a bool variable. That relies on the dummy/mock contract
being a new contract (in construction would set to false). Change
it to check "event" emitted instead so the code can be more reusable.

Also, adds some comments for the test only setters to make things clear.

The commit is based on the comment in following PR:
- PR: https://github.com/omisego/plasma-contracts/pull/127