fix: alter visibility for PriorityQueue(Lib)
Previously PriorityQueueLib was using 'public' instead of 'internal'.
This would make a 'delegetecall' instead of a 'jump' under the hood.
And would be more expensive on gas. It was intended to use 'public'
as previously it hits deploy gas cost limit. However, new abstract
design splits contracts into multiple ones, there is a potential to
not have such concern.

Also, changes 'public' to 'external' for PriorityQueue as it is not
needed to be public and external sometimes can safe a bit gas if input
is complex.

- PR: https://github.com/omisego/plasma-contracts/pull/137
- previous commit: https://github.com/omisego/plasma-contracts/commit/e63fc5e689486ef6600012517550c7fe79f1fc18