refactor: merge PriorityQueue and Lib into one
The PriorityQueueLib was originally moved out to lib to enable
separate deployment to cut down the gas cost. With new abstract
layer design contracts by default would not be as large as original.
Quite likely we are fine with deployment gas. Also, we can fallback to
use "init()" function to help on that as well.

This commit merges the lib back to decrease code complexity. Also add
some missing test on the "onlyOwner" part and remove unused var in struct.

PR: https://github.com/omisego/plasma-contracts/pull/137