fix: use 'exitable' timestamp as enqueue priority
This commit fixs the bug of using utxo_pos as priority. It should be the
exitable timestamp. See 'ExitableTimestamp' for timestamp logic.

Also, we change the timestamp to be using uint64 instead of uint256 to be able to pack
data into one field. The current standard timestamp is int32, but it would end at 2038,
which is not that far from now. Use 64 bits instead as most new 64 machines uses 64 bits.

-Year 2038 problem: https://en.wikipedia.org/wiki/Year_2038_problem
-GH issue: https://git.io/fj1we
-PR: https://github.com/omisego/plasma-contracts/pull/177