feat: add flag output spent functions
We need to flag output spent when exits are processed. We would
rely on this to resolve interaction between standard exit and
in-flight exit. Also, we would need this to solve the previous
problem that operator can steal fund.

This commit adds both batch and single api to flag. Batch api would
be used in in flight exit and single api for standard exit.

Meanwhile, as we moved some variables to public, which by default
comes with a getter function, remove the test for those getter.

- issue: https://github.com/omisego/plasma-contracts/issues/171
- operator steal fund: https://git.io/fj1rz
- PR: https://github.com/omisego/plasma-contracts/pull/179