fix: fix InFlightExit id miscalculation
We stated that the least 151 significant bits in exit id are reserved to
the tx's hash, but due to a bug it was no more that 105 bits of hash
used.

Shifting 256-bit left by 151 means that the 151 most significant bits
become zeroes and only 101 bits from the initial value are used (256 - 151 = 105).