chore: optimise with removing the need of storing 'exitableAt'
Since we are use 'exitableAt' in priority and we can parse that back,
we can stop storing 'exitableAt' to save some gas. Also, as the original
ExitModel.Exit contains only two fields, feel like worth just remove the
struct from 'enqueue' function and use the original field instead.

This commit optimise the storage for 'exitableAt' and change the api
signature of enqueue to remove struct. Also refactors the tests in
ExitGameController.