fix: add 'txPos' into exit priority consideration
It was a false statement that the exit priority is only based on
'exitableAt' in the previous commit. In fact, the 'exitableAt' only
provide granuanity of 'Block' level. We need to differenciate to 'Tx'
level. Otherwise earlier tx before invalid tx cannot have higher priority
when exiting.

This commit adds the 'txPos' into priority consideration and alter enqueue
api accordingly.