refactor: challengeSE to use router-controller
This commits refactors challengeStandardExit to use the router-controller
pattern. Meanwhile, there was a mysterious issue that when the controller library
is named 'PaymentChallengeStandardExitController', it keeps fail to link the lib
to the router contract in test. However, rename the lib solves the problem. As a
result, this commit also renames a bit on the convention of controller.