feat: replace OutputGuardParser with OutputGuardHandler
In previous implementation, we are bind to using output type 0 as some special
type that is not using outputGuard mechanism, which is hashing the preImages.
This commit tries to abstract that out and remove such assumption.

We introduce OutputGuardHandler that would take in OutputGuardModel.Object which
includes all essential data to verify a outputguard is valid or not. OuputGuardModel.Object
contains: guard, outputType, and preImage. Also, we use the OutputGuardHandler to retrieve
the exit target that is needed in ExitGame.