refactor: remove PaymentOutputGuardHandlerWrapper
If we returns a struct from an contract, it would be represented as Array actually.
Previously the Wrapper was there only to wrap the input args into a struct because it
keep failing to understand the js object as struct input. After turning it into Array
it works. Thus removing the unneccessary wrapper.