feat: remove outputGuardsForInputs from InFlightExit
Previously added a seperate list of outputGuards as it is needed
to get the essential data of 'exitTarget'. However, we agreed that
for inputs, it would be set on startIFE instead.

However, in challengeIFENonCanonical, we would need outputguard for current
spending condition interface. Put a empty bytes32 for now and would solve that
as the effort of https://github.com/omisego/plasma-contracts/issues/214.

discussion comment:
https://github.com/omisego/plasma-contracts/pull/222#discussion_r318611936