refactor: replace Constants with Protocol lib
Previously, it was using enum for PROTOCOL value, which is 0 and 1.
However, in the registry, it would default return value 0 if not registered.
This would lead to confusion when trying to retreive value from the registry.
To move to 1, 2 as protocol value, enum does not seem to be best way to do.
Thus move to a single Protocol lib and have value + getter function.