refactor: stub IFE data before nonCanonical challenge
It was sharing function with startIFE and calls a real startIFE and then
run the non-canonical challenge. However, as startIFE becomes more complex,
the setup for registries become complex if we want to keep the sharing test
file.

This commit remove the part that reuses and stub the test data directly
to the exit game contract to simplify the setup on registry and decrease
coupling between tests.