feat: add 'ercContract' to IExitProcessor interface
This commit add an extra var 'ercContract' to the exit processor interface. We need this in process
in-flight exit because one exitId will be reused in different exit queues with different erc contracts.
(PS. it is called 'token' mostly in the code for now, however, I think erc contract address represents
better)