269 new migration scripts (#298)
* fix: using wrong var for P2PCondition

* feat: setup networks

* feat: migration scripts until PaymentExitGame

Including PlasmaFramework, Vaults (ETH, ERC20), PaymentExitGame

* migration scripts

* updated migration scripts

* refactor: use configs file for constants in migration file

* chore: lower the gas for local network

As solidity optimizer has been turned on, we don't need such high gas now.

* fix: migration not starting with last migrated file

Add 'await' before 'deployer.deploy(Migrations);' to fix this.

* feat: use 'new' instead of 'deploy' for components on top of framework

Though it is convinient to use 'deploy', it does not support multiple instances of same contract yet.
As a result, for all the components we use 'new' instead. After registering to the framework, we can
get the contract address from framework by its public function.

see: https://github.com/trufflesuite/truffle/issues/237

* refactor: use single config instead of multiple ones

It is of best practice to put config in single file for simplicity.
https://codingsans.com/blog/node-config-best-practices

1. Merge previously two config files to one 'config.js'
2. Seperate config for different environments

* docs: add README on setting NODE_ENV for migration

* chore: truffle-hdwallet-provider -> truffle/hdwallet-provider

Old dependency is deprecated and should use the new one.

README: https://git.io/Jeng2

* fix: failed to get address without waiting

Previously the script uses PlsamaFramework.address without waiting for 'PlsamaFramework.deployed()'.
During the run of coverage it fails.

* chore: add coverage folder to gitignore

* fix: add and use maintainer address

Previously "authority address" is misused in most of the places, replace the wrong one
to the "maintainer address".

For detail of addresses: https://git.io/JenaT

* chore: lock down openzeppelin-solidity version to 0.2.3

We are importing openzeppelin solidity contracts in our contract codes. We should lock down as that
would be part of audit as well. We manually upgrade the version if we need it.

* fix: changes needed to rebase to master