Update PaymentProcessStandardExit.sol
Is there a comma needed in these lines? 
* @dev emits ExitOmitted event if the exit is omitted.
So it will look like this:  @dev emits ExitOmitted event, if the exit is omitted.  ADD A COMMA?
Also
ExitFinalized event if the exit is processed and funds are withdrawn
So it will look like this: 
ExitFinalized event, if the exit is processed and funds are withdrawn

This: @param exitMap The storage of all standard exit data.
Or this: @param exitMap Stores all standard exit data.