Update PaymentPiggybackInFlightExit.sol
Not clear what this line means: "Exit Target for outputs is set in piggyback instead of start in-flight exit."
Is it this? "Exit Target for outputs is set in the piggyback instead of when starting an in-flight exit."

Also find this in several files: The storage of all in-flight exit data.
Should it be: "Stores all in-flight exit data."?