242 separate maintainer address (#333)
* refactor: Replace Operated to OnlyFromAddress

We want to replace the word 'operator' to 'maintainer' in our code base as we have changed the
actors a bit (maintainer + authority). Also, we want to decouple maintainer with the deployer,
so we do not need to expose private key for maintainer during migration script running stage.

see:
- https://github.com/omisego/plasma-contracts/issues/242
- https://github.com/omisego/plasma-contracts/issues/304

* fix: tests for the refactor for Operated

* feat: replace initAuthority with activateChildChain

We want to store the authority address on deployment of contract to avoid
frontrunning the initAuthority method. Meanwhile, child-chain relies on
nonce to protect re-org. Thus we have this new 'activateChildChain' function
to be able to increase the nonce and block any block deposit before it is activated.

see: https://github.com/omisego/plasma-contracts/issues/304#issuecomment-538285330

* fix: inject authority and maintainer to framework in pytest

* refactor: use abstract getter for maintainer instead

Instead of dependency inject via constructor which leads to copying the storage, use a asbtract
getter function to get the maintainer address instead.

see: https://git.io/Je8is

* docs: docgen with the change of maintainer/authority address

* refactor: use onlyFrom(framework) in priority queue

To enhance the readibility that the owner of priority queue is expected to be PlsmaFramework.

* docs: docgen after change for priority queue

* refactor: use framework.getMaintainer() in Router

* docs: run docgen

* fix: grammar and typo from CR

Co-Authored-By: Kevin Sullivan <4653170+kevsul@users.noreply.github.com>

* refactor: use framewor.getMaintainer() instead

* docs: run docgen