feat: remove setAuthority function
That function was introduced to add the flexibility to move toward mechanism reorg protection
in the future. However, after discuss with the blockchain service team (elixir-omg) that feature
is not really on the priority list of implementing.

Also, auditor has sensed some issue with current way of updating authority. As a result of combination,
should just remove the function that we might not even needed to simply things.

issue: https://github.com/omisego/plasma-contracts/issues/403