feat: bump quarantined period for vault/depositVerifier to 2 minExitPeriod
An invalid vault or deposit verifier can cause user need to IFE if the tx is after the tx for
register vault or setDepositVerifier. This would make the default action different from registering
invalid ExitGame which is to start SE within a minExitPeriod.

For more detail, see the issue: https://github.com/omisego/plasma-contracts/issues/412