fix: ignore empty inputs/outpus in IFE data
Bring back the amount > 0 check. IFE data would be initiated with fixed size array for inputs/outputs.
As a result, we in fact do not want to care empty slots.

Also, change the unit test to be using real empty outputId in the empty slots to make the test more real.