feat: explicit revert when piggyback empty input/output slot
The IFE exit data is initiated with fixed size array for inputs/outputs. Thus, empty slot would be
default with empty value. This commit make piggyback fail explicitly when the input/output is empty.

Also, in processIFE, we treat 'isPiggybacked' check covering the 'isEmpty' check, as a result it is important
to make sure it fails to piggyback when empty.