feat: save IFE output owner when startIFE instead
PaymentExitGame was implemented to be able to support our restricted custody DEX design. As a result, the output guard preimage for an
in-flight exit tx output might not be availible. Without the preimage, it was not possible to know the where this output is exiting fund to.

However, we have decided to reduce the scope of Payment Exit Game and remove the DEX supporting feature such as output guard mechanism to simplify
the code base for initial launch. Given this, it becomes not necessary to save output owner (exitTarget) data in piggyback. It can be done in startIFE.