refactor: simplify tx finalization (#503)
* feat: MoreVpFinalization library as simplification

This commit adds a new lib MoreVpFinalization to handle the simplified version of
tx finalization verification logic. The plan is to replace all exisiting usage of
TxFinalizationVerifier and retire that afterward.

related issue: https://github.com/omisego/plasma-contracts/issues/471

* refactor: migrate startSE to MoreVpFinalization lib

* refactor: migrate challengeSE to MoreVpFinalization

* refactor: migrate startIFE to MoreVpFinalization

* refactor: migrate challengeIFENonCanonical to MoreVpFinalization

* refactor: check MoreVP protocol inside MoreVpFinalization

This commit also update codes that has migrated to MoreVpFinalization already

* refactor: migrate challengeIFEInputSPent to MoreVpFinalization

1. Migrate challengeIFEInputSpent to new finalization lib
2. Use assert on 'isProtocolFinalized' as MoreVP only checks the existence of tx. An empty tx would always fail earlier.

* refactor: migrate challengeIFEOutputSpent to MoreVpFinalization

1. Fix a bug that original code is checking in-flight tx standard finalized. That is already checked during startIFE. Should be checking
the challenging tx is protocol finalized instead.
2. migrate to new lib

* refactor: remove old TxFinalization

1. Remove the code
1. Remove the args from PaymentExitGame
1. Update migration script and tests for those
1. Comment out original test and prepare to move it to MoreVpFinalization tests later

* refactor: remove unused args from e2e tests

* fix: add standard finalize check back for IFE tx when challengeOutputSpent

Previous commit removed it wrongly. When IFE tx is spend, it must be included in a block (standard finalized).
This commit add the check and test back.

* refactor: remove optional args from spending condition interface

Since that is not used in Payment V1, as a result, remove unused code. Also, for future proofing, extra args can still
be passed as 'witness', just the witness need to be encoded in a way spending condition implementation can understand.

* fix: python test remove TxFinalizationVerifier

* refactor: original finalization lib test to MoreVpFinalization tests

* chore: add test coverage on protocol getter

Coverage droppned while refactoring. As currently the value of 'MVP()' is not used.
We only check protocol != More_VP() now. This commit add test to bring up the coverage.

* fix: apply suggestions from code review

Co-Authored-By: Kevin Sullivan <4653170+kevsul@users.noreply.github.com>

* refactor: remove wrong comments

* docs: npm run docgen