fee tx extension (#518)
* feat: add fee contracts

We are adding fee feature with Payment V1. This commit adds FeeExitGame and FeeClaimOutputToPaymentTxCondition contracts.

* feat: add fee contracts to migration script and config

- Add fee tx type and output type to config.js
- Add new file 6_deploy_and_register_fee_exit_game
- deploy FeeClaimOutputToPaymentTxCondition together with 5_deploy_and_register_payment_exit_game
- Bump the migration script numbers after the file 6_deploy_and_register_fee_exit_game

* refactor: remove unused var in standard exit e2e test

* feat: bump initialImmuneExitGames to 2 in config

Adding Fee Exit Game as exception.

* feat: add feeBlockNum output type

* test: fee claim e2e tests

* fix: typo in paymentToPaymentSpendingCondition

* test: FeeClaimOutputToPaymentTxCondition unit tests

* fix: change field name in paymentEip712

The fourth field name is changed from metaData to txData. However, paymentEip712 is still using metaData.
This change it to be using txData instead. Also, remove the auto fill of EMPTY_BYTE32 as it is promised for
the payment tx to have meta data now.

* test: make fee claim e2e test more similar to real situation

The test was written with invalid block that the fee tx claims the fee out of air. This commit adds the flow that
the fee is indeed collecting from other txs within the block.

* fix: for PR comments

* feat: remove fee blockNum output

Instead of relying on an extra output to hold the "uniqueness salt" data (blockNum or nonce), use
the fourth field of Generic Transaction instead.

see: https://github.com/omisego/feat-fee/pull/47

* fix: rebase to latest Generic Transaction

Was making changes while the PR for Generic Transaction is not merged yet.

* feat: replace blockNum with nonce in FeeTx txData

* docs: add fee doc and update payment doc

We increase the scope of payment v1 to be with fee tx as input. This commit update the docs with the changes.

* fix: apply suggestions from code review

Co-Authored-By: Kevin Sullivan <4653170+kevsul@users.noreply.github.com>