fix: mekle contract checks the index size (#547)
* fix: mekle contract checks the index size

Merkle contract did not check whether index exceeds the max amount with the proof size. This can lead to issue where
given different indexes that exceeds the size, it can still pass the check due to the implementation of for loop.

see issue for more details:
https://github.com/omisego/plasma-contracts/issues/546

* fix: mekle js helper with correct height

merkle.js was calculating with the wrong height when the number of leaves is of exponential of 2. (eg. 4, 8, 16...)

* fix: code review comment