test: integ test for challenge IFE input spend for all output indexes
This is an effort to add tests to cover challenge IFE input spent on all output indexes.
We found a bug that we are using wrong indexes in contract code.

But, this test end up does not break the existing contract as although it is attracting wrong output, we are only
using the output type with that output. With current system, we only have 1 output type for payment tx. As a result
it failed to break with this test..

https://github.com/omisego/security-issues/issues/15