chore: output gas report to file instead of stdout
This commit change the gas report output to a file. However, even it is state in the document that it should be
RST format, it does not read that well. It read best in terminal somehow. Thus use `cat` command to print the result in circleCI.

Also, send the gasReport.rst to circle ci artifect