chore: check api token before submission + save to artifact (#613)
* chore: check api token for coverage and gas report

If using fork or private mirror repo, the CI job would not be able to submit with the token accordingly.
This commit omit the submission to codecheck and coverall if without token.

* chore: save coverage file to artifect

* chore: save gas report to artifect

* chore: output gas report to file instead of stdout

This commit change the gas report output to a file. However, even it is state in the document that it should be
RST format, it does not read that well. It read best in terminal somehow. Thus use `cat` command to print the result in circleCI.

Also, send the gasReport.rst to circle ci artifect