fix: flaky python test for IFE process
The flaky python test setup was doomed to be flaky as there is no promise on the priority of 2 IFEs.
In fact the setup is more like broken where a non-canonical IFE should be challenged instead of processed as canonical.
This commit change the test setup and rename the test to a non-canonical tx challenge scenario.

see: https://github.com/omgnetwork/plasma-contracts/issues/606#issuecomment-678871977